Create a folder named Publishers and move the single publisher to it
