+++ components/email-service/test/Altinn.Notifications.Email.Tests/Email.Integrations/ServiceCollectionExtensionsTests.cs
@@ -1,8 +1,7 @@
 ﻿using Altinn.Notifications.Email.Core.Dependencies;
 using Altinn.Notifications.Email.Integrations.Configuration;
 using Altinn.Notifications.Email.Integrations.Consumers;
-using Altinn.Notifications.Email.Integrations.Producers;
-
+using Altinn.Notifications.Email.Integrations.Publishers;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
